feat: OpenApiProvider 구현 및 지수 데이터/정보 관련 DTO 추가 ApiResponse 타입 수정

+++ src/main/java/com/team3/findex/common/openapi/IndexApiResponse.java
@@ -3,6 +3,7 @@ package com.team3.findex.common.openapi;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.Getter;
 
 /**
@@ -34,22 +35,22 @@ public class IndexApiResponse {
   private Integer itemsCount;
 
   @JsonProperty("clpr")
-  private Double closingPrice;
+  private BigDecimal closingPrice;
 
   @JsonProperty("vs")
-  private Double variationSign;
+  private BigDecimal variationSign;
 
   @JsonProperty("fltRt")
-  private Double changeRate;
+  private BigDecimal changeRate;
 
   @JsonProperty("mkp")
-  private Double marketPrice;
+  private BigDecimal marketPrice;
 
   @JsonProperty("hipr")
-  private Double highPrice;
+  private BigDecimal highPrice;
 
   @JsonProperty("lopr")
-  private Double lowPrice;
+  private BigDecimal lowPrice;
 
   @JsonProperty("trqu")
   private Long tradeVolume;
